==> notebooks/01_STOPPD_terminal.Rmd <==
---
title: "01_STOPPD_terminal"
author: "Navona"
written: January 18, 2018  
last run: Sys.Date()
output: html_document
---

#This script pulls in and cleans up the naming of STOPPD scans as they exist in the Kimel lab file system. At earlier stages, this script helped us identify naming errors in the file system (all have since been fixed).

#Purpose: The contents of the file system will, in other scripts, be checked against (1) the scans we have in XNAT, to ensure that there are no discrepancies between these databases, and also against (2) our subject inclusion list.

```{r setup}
library('stringi')
library('stringr')
library('plyr')
library('tidyr')
```

```{r import_data}

#import spreadsheet ('ls' of file system)
terminal <- read.csv('../data/stoppd_NiiFolderContents_2018-01-25.csv', header = TRUE, stringsAsFactors = FALSE)

```

```{r data_cleaning}

#make a new column for site component of ID
terminal$site <- str_sub(terminal$scan_id, 8, 10)

#cut out study and site component from ID (first 11 characters)
terminal$scan_id <- substring(terminal$scan_id, 12)

#make a new column for session component of ID
terminal$session <- str_sub(terminal$scan_id, -2)

#cut out session information from ID (last 3 characters)
terminal$scan_id <- stri_sub(terminal$scan_id, 1, -4)

#make a new column that captures alphabetic component of ID ('R')
terminal$contains_R <- grepl('R', terminal$scan_id, fixed=TRUE) #36 participants

#cut out the 'R' in some participant IDs (indicates repeat for controls)
terminal$scan_id <- gsub("[R]", "", terminal$scan_id)

#make a 'group' column to capture case vs. control information
terminal$group <- stri_sub(terminal$scan_id, 2, 2) #note: 1 or 2 is patient, 6 is control

#for clarity, change values in 'group' column to labels for clarity
terminal$group[terminal$group == 1] <- "patient"
terminal$group[terminal$group == 2] <- "patient"
terminal$group[terminal$group == 6] <- "control"

#make a variable that combines unique ID and session number
terminal$id_session <- paste(terminal$scan_id, '_', terminal$session, sep='')

#write csv
write.csv(terminal, '../generated_csvs/terminal_clean_2018-01-25.csv', row.names=FALSE)

#cleanup
rm(terminal)

```
